Fold directive-on-input-field tests into existing test classes
